requirements.txt
alembic==1.7.7
aniso8601==9.0.1
astroid==2.15.6
attrs==21.4.0
black==23.9.1
cfgv==3.4.0
click==8.1.3
colorama==0.4.4
dill==0.3.7
distlib==0.3.7
filelock==3.12.4
flake8==6.1.0
Flask==2.1.2
flask-accepts==0.18.4
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
identify==2.5.28
importlib-metadata==4.11.4
importlib-resources==5.7.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
lazy-object-proxy==1.9.0
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.19.0
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.0
pathspec==0.11.2
platformdirs==3.10.0
pre-commit==3.4.0
psycopg2-binary==2.9.5
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.6.0
pylint==2.17.5
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==1.4.36
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
virtualenv==20.24.5
Werkzeug==2.0.3
wrapt==1.15.0
zipp==3.8.0
pytest==7.4.2
pytest-cov==4.1.0
coverage==7.3.1
factory_boy==3.2.1